requirements.txt
torch==2.0.1
torchvision==0.15.2
xgboost==2.0.3
numpy==1.24.3
opencv-python==4.8.0.74
pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
werkzeug==2.3.7
tqdm==4.65.0
joblib==1.3.2
